+++ ruff.toml
@@ -5,7 +5,7 @@ extend-exclude = [
     "env/",
     ".env",
     ".env.*",
-    "migrations"
+    "alembic"
 ]
 
 [lint]
