+++ src/main.py
@@ -2,7 +2,7 @@
 from fastapi import FastAPI
 
 from src.routers.auth_routers import router as auth_router
-from src.routers.auth_user_routes import router as main_router
+from src.routers.user_routes import router as main_router
 
 
 app = FastAPI()
